Refactor warn.go
