+++ warn/BUILD.bazel
@@ -5,7 +5,12 @@ go_library(
     srcs = [
         "types.go",
         "warn.go",
-        "warn_bazel.go",
+        "warn_api.go",
+        "warn_appearance.go",
+        "warn_execution.go",
+        "warn_misc.go",
+        "warn_operation.go",
+        "warn_syntax.go",
     ],
     importpath = "github.com/bazelbuild/buildtools/warn",
     visibility = ["//visibility:public"],
@@ -22,7 +27,12 @@ go_test(
     size = "small",
     srcs = [
         "types_test.go",
-        "warn_bazel_test.go",
+        "warn_api_test.go",
+        "warn_appearance_test.go",
+        "warn_execution_test.go",
+        "warn_misc_test.go",
+        "warn_operation_test.go",
+        "warn_syntax_test.go",
         "warn_test.go",
     ],
     embed = [":go_default_library"],
